Change addgif
Readded the database in the addgif command. The bot will store a description
of the GIF, which is optional, and it will also store the filename. Continuing
to use the database to store some GIF information so it can keep track of IDs
and descriptions to make listing, deleting, and viewing GIFs a little more
user friendly.