Add CustomGIF changes to commands
All commands have been modified to handle the new CustomGIFs system, but
there are some issues with speed which is timing out the commands
occasionally. This appears to only happen when a command is run in which
a GIF will be loaded. If the command does not have a stored GIF to
consider, it will work properly.