Add new boost module
This module adds 3 commands. /boost setchannel, setimage, and setmessage.
When a channel is set, the bot will send a custom message when a user
boosts the server. The embed can be modified with setmessage and setimage.
setimage and setmessage cannot be executed unless a channel is set with
setchannel, and the alerts can be disabled by unsetting the channel.